fix(vst3): cast COINIT_APARTMENTTHREADED i32 → u32 pour CoInitializeEx
windows-sys 0.59 expose les constantes COINIT_* en i32 mais la signature
CoInitializeEx attend dwCoInit: u32. Cast direct au call site.

+++ jamodio-vst3-host/src/editor.rs
@@ -439,7 +439,8 @@ fn editor_thread_main(data: EditorThreadData) {
     // MFC framework, etc.) qui exigent un thread STA. Sans CoInitializeEx,
     // ces appels deadlock silencieusement ou la window reste blanche +
     // "NOT RESPONDING".
-    let com_init = unsafe { CoInitializeEx(std::ptr::null(), COINIT_APARTMENTTHREADED) };
+    let com_init =
+        unsafe { CoInitializeEx(std::ptr::null(), COINIT_APARTMENTTHREADED as u32) };
     tracing::info!(
         target: "jamodio::vst3::editor",
         hresult = format!("0x{com_init:08X}"),
